fix(query-editor): unique input field ids per query (#538)
Introduced a regression in the latest version.

Since the same label is used now, when having multiple queries and by clicking any switch on the `Query X` would always change `Query A` instead.

To fix this we need to assign unique ids for each labelled field (based on query refId).